myorg/requirements.txt
Django==4.2.30
djangorestframework==3.17.1
drf-yasg==1.21.15
dj-database-url==3.1.2
gunicorn==25.3.0
psycopg2-binary==2.9.12
whitenoise==6.12.0
